+++ controllers/auth/indieauth/client.go
@@ -6,16 +6,16 @@ import (
 	"io"
 	"net/http"
 
-	"github.com/owncast/owncast/auth"
 	ia "github.com/owncast/owncast/auth/indieauth"
 	"github.com/owncast/owncast/controllers"
 	"github.com/owncast/owncast/core/chat"
-	"github.com/owncast/owncast/core/user"
+	"github.com/owncast/owncast/models"
+	"github.com/owncast/owncast/persistence/userrepository"
 	log "github.com/sirupsen/logrus"
 )
 
 // StartAuthFlow will begin the IndieAuth flow for the current user.
-func StartAuthFlow(u user.User, w http.ResponseWriter, r *http.Request) {
+func StartAuthFlow(u models.User, w http.ResponseWriter, r *http.Request) {
 	type request struct {
 		AuthHost string `json:"authHost"`
 	}
@@ -63,15 +63,17 @@ func HandleRedirect(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	userRepository := userrepository.Get()
+
 	// Check if a user with this auth already exists, if so, log them in.
-	if u := auth.GetUserByAuth(response.Me, auth.IndieAuth); u != nil {
+	if u := userRepository.GetUserByAuth(response.Me, models.IndieAuth); u != nil {
 		// Handle existing auth.
 		log.Debugln("user with provided indieauth already exists, logging them in")
 
 		// Update the current user's access token to point to the existing user id.
 		accessToken := request.CurrentAccessToken
 		userID := u.ID
-		if err := user.SetAccessTokenToOwner(accessToken, userID); err != nil {
+		if err := userRepository.SetAccessTokenToOwner(accessToken, userID); err != nil {
 			controllers.WriteSimpleResponse(w, false, err.Error())
 			return
 		}
@@ -90,14 +92,14 @@ func HandleRedirect(w http.ResponseWriter, r *http.Request) {
 
 	// Otherwise, save this as new auth.
 	log.Debug("indieauth token does not already exist, saving it as a new one for the current user")
-	if err := auth.AddAuth(request.UserID, response.Me, auth.IndieAuth); err != nil {
+	if err := userRepository.AddAuth(request.UserID, response.Me, models.IndieAuth); err != nil {
 		controllers.WriteSimpleResponse(w, false, err.Error())
 		return
 	}
 
 	// Update the current user's authenticated flag so we can show it in
 	// the chat UI.
-	if err := user.SetUserAsAuthenticated(request.UserID); err != nil {
+	if err := userRepository.SetUserAsAuthenticated(request.UserID); err != nil {
 		log.Errorln(err)
 	}
