+++ controllers/notifications.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"net/http"
 
-	"github.com/owncast/owncast/core/user"
+	"github.com/owncast/owncast/models"
 	"github.com/owncast/owncast/notifications"
 
 	"github.com/owncast/owncast/utils"
@@ -14,8 +14,8 @@ import (
 
 // RegisterForLiveNotifications will register a channel + destination to be
 // notified when a stream goes live.
-func RegisterForLiveNotifications(u user.User, w http.ResponseWriter, r *http.Request) {
-	if r.Method != POST {
+func RegisterForLiveNotifications(u models.User, w http.ResponseWriter, r *http.Request) {
+	if r.Method != http.MethodPost {
 		WriteSimpleResponse(w, false, r.Method+" not supported")
 		return
 	}
